+++ Pod/Classes/SEGAppboyIntegrationEndpointDelegate.h
@@ -1,0 +1,10 @@
+#import "Foundation/Foundation.h"
+#import "ABKAppboyEndpointDelegate.h"
+
+@interface SEGAppboyIntegrationEndpointDelegate : NSObject <ABKAppboyEndpointDelegate>
+
+@property NSString *customEndpoint;
+
+- (instancetype)initWithCustomEndpoint:(NSString *)customEndpoint;
+
+@end